Get repository to build with GCC 11
See #594

+++ libc/testlib/strcaseequals.c
@@ -21,7 +21,7 @@
 #include "libc/testlib/testlib.h"
 
 bool testlib_strcaseequals(size_t cw, const void *s1, const void *s2) {
-  return testlib_strncaseequals(cw, s1, s2, SIZE_MAX);
+  return testlib_strncaseequals(cw, s1, s2, LONG_MAX);
 }
 
 bool testlib_strncaseequals(size_t cw, const void *s1, const void *s2,
